requirements.txt
aiogram==2.25.2
dotenv
pillow
python-docx
wikipedia